+++ src/types/models.type.d.ts
@@ -1,0 +1,2 @@
+
+export type PlansTypes = 'free' | 'basic' | 'premium'
